identified a flaw. base and target should be flipped for the historical prices at timestamps for the coinbase exchange.
should put timestamp prices in a new table on db, with a field for which exchange its for.
they should be kept separate from the aggregated prices so that we can use prices that are as close to exact as possible when doing math on each transaction.
maybe don't do the extra network requests if the transaction contains all the info necessary without an additional price request